requirements.txt
absl-py==1.2.0
astor==0.8.1
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
dlib==19.24.0
gast==0.5.3
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
imageio==2.22.0
importlib-metadata==5.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.6.3
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.0
pyparsing==3.0.9
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==2.0.1
tifffile==2021.11.2
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.64.1
typing_extensions==4.3.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
